refactor: centralize generated Solidity bindings
