+++ .github/workflows/e2e.yml
@@ -29,11 +29,15 @@ jobs:
           filters: |
             lint:
               - "e2e/**"
+              - "gen/go/solidity-abi/**"
+              - "go.work"
               - ".golangci.yml"
               - ".github/workflows/e2e.yml"
             code:
               - "link/**"
               - "e2e/**"
+              - "gen/go/solidity-abi/**"
+              - "go.work"
               - "proto/**"
               - "Makefile"
               - ".github/workflows/e2e.yml"
@@ -47,11 +51,17 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: e2e/go.mod
-          cache-dependency-path: e2e/go.sum
+          cache-dependency-path: |
+            e2e/go.sum
+            gen/go/solidity-abi/go.sum
       - uses: golangci/golangci-lint-action@ba0d7d2ec06a0ea1cb5fa41b2e4a3ab91d21278a # v9.3.0
         with:
           version: v2.12.2
           working-directory: e2e
+      - uses: golangci/golangci-lint-action@ba0d7d2ec06a0ea1cb5fa41b2e4a3ab91d21278a # v9.3.0
+        with:
+          version: v2.12.2
+          working-directory: gen/go/solidity-abi
 
   test:
     needs: changes
@@ -65,6 +75,7 @@ jobs:
           cache-dependency-path: |
             link/go.sum
             e2e/go.sum
+            gen/go/solidity-abi/go.sum
       - run: make build-link
       - run: docker info
       - run: go test -count=1 ./internal/... ./cmd/...
@@ -103,6 +114,7 @@ jobs:
           cache-dependency-path: |
             link/go.sum
             e2e/go.sum
+            gen/go/solidity-abi/go.sum
       - run: docker info
       - run: make test-e2e
 
@@ -113,6 +125,8 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: e2e/go.mod
-          cache-dependency-path: e2e/go.sum
+          cache-dependency-path: |
+            e2e/go.sum
+            gen/go/solidity-abi/go.sum
       - run: docker info
       - run: make check-e2e-matrix
