+++ .github/workflows/ibc-link-build.yml
@@ -54,7 +54,9 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: link/go.mod
-          cache-dependency-path: link/go.sum
+          cache-dependency-path: |
+            link/go.sum
+            gen/go/solidity-abi/go.sum
       - name: Build
         env:
           CGO_ENABLED: "0"
