+++ .github/workflows/ibc-link-ci.yml
@@ -38,6 +38,8 @@ jobs:
           filters: |
             link:
               - 'link/**'
+              - 'gen/go/solidity-abi/**'
+              - 'go.work'
               - '.golangci.yml'
               - '.github/workflows/ibc-link-ci.yml'
             sql:
@@ -61,7 +63,9 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: link/go.mod
-          cache-dependency-path: link/go.sum
+          cache-dependency-path: |
+            link/go.sum
+            gen/go/solidity-abi/go.sum
       - uses: golangci/golangci-lint-action@ba0d7d2ec06a0ea1cb5fa41b2e4a3ab91d21278a # v9
         with:
           version: v2.12.2
@@ -80,7 +84,9 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: link/go.mod
-          cache-dependency-path: link/go.sum
+          cache-dependency-path: |
+            link/go.sum
+            gen/go/solidity-abi/go.sum
       - name: Download dependencies
         run: go mod download
       - name: Build
@@ -102,7 +108,9 @@ jobs:
       - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7
         with:
           go-version-file: link/go.mod
-          cache-dependency-path: link/go.sum
+          cache-dependency-path: |
+            link/go.sum
+            gen/go/solidity-abi/go.sum
       - name: Test
         run: make test
