+++ e2e/internal/e2etest/traffic_ift.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/cosmos/ibc/e2e/internal/harness/chain/evm"
 	"github.com/cosmos/ibc/e2e/internal/harness/environment"
-	"github.com/cosmos/ibc/e2e/internal/harness/environment/solidityibc/iftbatchtransfershim"
+	"github.com/cosmos/ibc/gen/go/solidity-abi/iftbatchtransfershim"
 )
 
 var (
