+++ e2e/internal/harness/environment/solidityibc/setup.go
@@ -24,8 +24,8 @@ import (
 	"github.com/ethereum/go-ethereum/ethclient"
 
 	"github.com/cosmos/ibc/e2e/internal/harness/chain/evm"
-	"github.com/cosmos/ibc/e2e/internal/harness/environment/solidityibc/accessmanager"
-	"github.com/cosmos/ibc/e2e/internal/harness/environment/solidityibc/escrow"
+	"github.com/cosmos/ibc/gen/go/solidity-abi/accessmanager"
+	"github.com/cosmos/ibc/gen/go/solidity-abi/escrow"
 )
 
 type contractBackend interface {
