+++ gen/go/solidity-abi/accessmanager/contract.go
