+++ gen/go/solidity-abi/escrow/contract.go
