+++ gen/go/solidity-abi/iftbatchtransfershim/contract.go
